✨feat: implement download image function
