+++ app/@modal/(.)users/[key]/page.tsx
@@ -17,6 +17,7 @@ export default async function ImageModal({
     <Modal>
       <Image
         src={`data:${contentType};base64,${base64}`}
+        id={key}
         width={100}
         height={100}
         className="w-full"
