+++ app/components/step.tsx
@@ -1,11 +1,13 @@
+import type { Steps } from '@/app/interfaces/steps'
+
 export function Step({
-  step,
+  steps,
   targetStep,
-}: { step: string[]; targetStep: number }) {
+}: { steps: Steps; targetStep: number }) {
   return (
     <>
       <ul className="steps">
-        {step.map((step, index) => (
+        {steps?.map((step, index) => (
           <li
             key={step}
             className={`step${index <= targetStep ? ' step-primary' : ''}`}
@@ -14,7 +16,7 @@ export function Step({
           </li>
         ))}
       </ul>
-      <h1 className="font-semibold text-2xl">{step[targetStep]}</h1>
+      <h1 className="font-semibold text-2xl">{steps?.[targetStep]}</h1>
     </>
   )
 }
