+++ app/interfaces/steps.ts
@@ -1,0 +1,1 @@
+export type Steps = string[]
