+++ app/register/page.tsx
@@ -1,12 +1,13 @@
 import { Step } from '@/app/components/step'
+import type { Steps } from '@/app/interfaces/steps'
 import { ProfileForm } from './components/profileForm'
 
-export const STEP = ['必要情報の入力', '入力確認', '完了']
+export const STEPS: Steps = ['必要情報の入力', '入力確認', '完了'] as const
 
 export default function Register() {
   return (
     <>
-      <Step step={STEP} targetStep={0} />
+      <Step steps={STEPS} targetStep={0} />
       <ProfileForm />
     </>
   )
