+++ app/users/[key]/page.tsx
@@ -1,3 +1,4 @@
+import { DownloadBtn } from '@/app/components/downloadBtn'
 import { getImage } from '@/app/lib/getImage'
 import { ArrowUturnLeftIcon } from '@heroicons/react/24/solid'
 import Image from 'next/image'
@@ -19,10 +20,12 @@ export default async function ImagePage({
       <h1 className="font-semibold text-2xl">{key}</h1>
       <Image
         src={`data:${contentType};base64,${base64}`}
+        id={key}
         width={300}
         height={300}
         alt={key}
       />
+      <DownloadBtn />
       <Link href="/users" className="btn">
         <ArrowUturnLeftIcon className="size-6" />
         表に戻る
